+++ Sources/Cachebay/Compiler/CachePlanLiteral.swift
@@ -65,6 +65,8 @@ public extension PlanField {
             connectionFilters: connectionFilters,
             connectionMode: connectionMode,
             pageArgs: pageArgs,
+            skipIf: nil,
+            includeIf: nil,
             selId: selId
         )
     }